+++ src/main/java/org/folio/ld/fingerprint/service/FingerprintServiceImpl.java
@@ -36,10 +36,16 @@ public class FingerprintServiceImpl implements FingerprintService {
   public String fingerprint(Resource resource) {
     var matchedRule = rules.getRules()
       .stream()
-      .filter(rule -> resource.getTypeNames().equals(rule.types()))
+      .filter(rule -> typesMatch(resource, rule))
       .findFirst();
     var fingerprintMap = getFingerprint(resource, matchedRule.orElse(null));
     return mapper.writeValueAsString(fingerprintMap);
+  }
+
+  private static boolean typesMatch(Resource resource, FingerprintRule rule) {
+    return Boolean.TRUE == rule.partialTypesMatch()
+      ? resource.getTypeNames().containsAll(rule.types())
+      : resource.getTypeNames().equals(rule.types());
   }
 
   private List<FingerprintEntry> getFingerprint(Resource resource, FingerprintRule rule) {