+++ src/test/java/org/folio/ld/fingerprint/test/TestUtil.java
@@ -114,6 +114,7 @@ import static org.folio.ld.dictionary.PropertyDictionary.TYPE_OF_REPORT;
 import static org.folio.ld.dictionary.PropertyDictionary.VARIANT_TYPE;
 import static org.folio.ld.dictionary.PropertyDictionary.WITH_NOTE;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.ANNOTATION;
+import static org.folio.ld.dictionary.ResourceTypeDictionary.BOOKS;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.CATEGORY;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.CATEGORY_SET;
 import static org.folio.ld.dictionary.ResourceTypeDictionary.CONCEPT;
@@ -328,7 +329,7 @@ public class TestUtil {
     pred2OutgoingResources.put(MEDIA, List.of(category("MEDIA")));
     pred2OutgoingResources.put(CARRIER, List.of(carrier));
     pred2OutgoingResources.put(COPYRIGHT, List.of(copyrightEvent));
-    pred2OutgoingResources.put(INSTANTIATES, List.of(work()));
+    pred2OutgoingResources.put(INSTANTIATES, List.of(work(BOOKS)));
 
     var instance = getResource(
       Map.ofEntries(
@@ -384,7 +385,7 @@ public class TestUtil {
     return instance.setLabel("Instance label");
   }
 
-  public static Resource work() {
+  public static Resource work(ResourceTypeDictionary subType) {
     var place = getResource(
       Map.of(
         NAME, List.of("United States"),
@@ -537,7 +538,7 @@ public class TestUtil {
         DATE_START, List.of("2023"),
         DATE_END, List.of("2024")
       ),
-      Set.of(WORK),
+      Set.of(WORK, subType),
       pred2OutgoingResources
     ).setLabel("Work: label");
   }