+++ NEWS.md
@@ -2,6 +2,7 @@
 - Add fingerprint rule for EXTENT resource [MODLD-390](https://folio-org.atlassian.net/browse/MODLD-390)
 - Mapping for MARC field 490 [MODLD-714](https://folio-org.atlassian.net/browse/MODLD-714)
 - "targetAudience", "geographicCoverage" links contain https instead of http [MODLD-695](https://folio-org.atlassian.net/browse/MODLD-695)
+- Partial types rule selection for WORK [MODLD-805](https://folio-org.atlassian.net/browse/MODLD-805)
 
 ## 1.0.1 (03-12-2025)
 - Initial release